Remove Redux and fetch from App to Orders.js
