+++ src/components/Orders/Orders.js
@@ -1,25 +1,59 @@
-import React from 'react';
+import React, { Component } from 'react';
 import './Orders.css';
+import { connect } from 'react-redux';
+import { bindActionCreators } from 'redux';
+import { setOrders } from '../../actions';
+import { getOrders } from '../../apiCalls';
 
-const Orders = props => {
-  const orderEls = props.orders.map(order => {
-    return (
-      <div className="order">
-        <h3>{order.name}</h3>
-        <ul className="ingredient-list">
-          {order.ingredients.map(ingredient => {
-            return <li>{ingredient}</li>
-          })}
-        </ul>
-      </div>
-    )
-  });
+export class Orders extends Component {
+  constructor(props) {
+    super();
+    // this.props = props;
+  }
 
-  return (
-    <section>
-      { orderEls.length ? orderEls : <p>No orders yet!</p> }
-    </section>
-  )
+  componentDidMount() {
+    getOrders()
+      .then(data => this.props.setOrders(data.orders))
+      .catch(err => console.error('Error fetching:', err));
+  }
+
+
+  orders = () => {
+    // const orderEls = this.props.orders
+    this.props.orders.map(order => {
+      return (
+        <div className="order">
+          <h3>{order.name}</h3>
+          <ul className="ingredient-list">
+            {order.ingredients.map(ingredient => {
+              return <li>{ingredient}</li>
+            })}
+          </ul>
+        </div>
+      )
+    });
+  }
+
+    render() {
+      return (
+        <section>
+          { this.props.orders.length ? this.orders : <p>No orders yet!</p> }
+        </section>
+      )
+    }
 }
 
-export default Orders;
+
+export const mapStateToProps = ({ orders }) => ({
+  orders
+});
+
+export const mapDispatchToProps = dispatch => (
+  bindActionCreators({
+    setOrders,
+  }, dispatch)
+);
+
+export default connect(mapStateToProps, mapDispatchToProps)(Orders);
+
+// export default Orders;
