+++ src/containers/App/App.js
@@ -1,23 +1,18 @@
 import React, { Component } from 'react';
 import './App.css';
-import { connect } from 'react-redux';
-import { bindActionCreators } from 'redux';
-import { setOrders } from '../../actions';
-import { getOrders } from '../../apiCalls';
+// import { connect } from 'react-redux';
+// import { bindActionCreators } from 'redux';
+// import { setOrders } from '../../actions';
+// import { getOrders } from '../../apiCalls';
 import Orders from '../../components/Orders/Orders';
 import OrderForm from '../../components/OrderForm/OrderForm';
 
 export class App extends Component {
   constructor(props) {
     super();
-    this.props = props;
   }
 
-  componentDidMount() {
-    getOrders()
-      .then(data => this.props.setOrders(data.orders))
-      .catch(err => console.error('Error fetching:', err));
-  }
+
 
   render() {
     return (
@@ -27,21 +22,12 @@ export class App extends Component {
           <OrderForm />
         </header>
         
-        <Orders orders={this.props.orders}/>
+        <Orders />
+
       </main>
     );
   }
 }
 
-export const mapStateToProps = ({ orders }) => ({
-  orders
-});
 
-export const mapDispatchToProps = dispatch => (
-  bindActionCreators({
-    setOrders,
-  }, dispatch)
-);
-
-export default connect(mapStateToProps, mapDispatchToProps)(App);
-
+export default App
